requirements.txt
selenium==4.11.2
webdriver-manager==4.0.0
Pillow==10.0.0